Fix PASSEDBYVALUE handling for uint8 type
Previously, we tried to parse pg_config.h to find out whether we need
to declare the uint8 type as PASSEDBYVALUE.  This turned out to be
very fragile on Red Hat-based systems, because the pg_config.h file is
in unusual places and split across several files.  The
non-PASSEDBYVALUE case is only applicable to 32-bit systems, which are
very rare nowadays.  So to simplify all this and make it more robust,
just hardcode PASSEDBYVALUE and thus drop support for 32-bit systems.

fixes #17